consolidate winui sample
