+++ Samples/WinUIDesktopSample/App.xaml.cs
