+++ Samples/WinUIDesktopSample/WinUIDesktopSample.exe.manifest
