--- WinUIDesktopSample/Net5_WinUI_Fixup.Targets
@@ -1,78 +1,0 @@
-<Project>
-
-  <!--Set properties expected by the real Microsoft.AppxPackage.Targets, in evaluating its own properties and items-->
-  <PropertyGroup>
-    <TargetPlatformIdentifier>Windows</TargetPlatformIdentifier>
-    <TargetPlatformMinVersion>10.0.18362.0</TargetPlatformMinVersion>
-    <TargetPlatformVersion>$(TargetPlatformMinVersion)</TargetPlatformVersion>
-    <WindowsSdkPath Condition="'$(WindowsSdkPath)' == ''">$([Microsoft.Build.Utilities.ToolLocationHelper]::GetPlatformSdkLocation("Windows", "10.0"))</WindowsSdkPath>
-    <PlatformXmlDir>$(WindowsSdkPath)\Platforms\UAP\$(TargetPlatformVersion)\</PlatformXmlDir>
-    <RealMsAppxPackageTargets>$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\AppxPackage\Microsoft.AppXPackage.Targets</RealMsAppxPackageTargets>
-  </PropertyGroup>
-
-  <Import Project="$(RealMsAppxPackageTargets)"/>
-
-  <!--Clear out properties expected by Microsoft.Net.SDK-->
-  <PropertyGroup>
-    <TargetPlatformIdentifier></TargetPlatformIdentifier>
-    <TargetPlatformVersion></TargetPlatformVersion>
-  </PropertyGroup>
-
-  <!--Hook every target in the real Microsoft.AppxPackage.Targets, setting and restoring needed properties-->
-  <Target Name="Before_GetSdkToolsPathsFromSdk" BeforeTargets="_GetSdkToolsPathsFromSdk">
-    <PropertyGroup>
-      <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
-      <TargetPlatformVersion>$(TargetPlatformMinVersion)</TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="After_GetSdkToolsPathsFromSdk" AfterTargets="_GetSdkToolsPathsFromSdk">
-    <PropertyGroup>
-      <TargetPlatformIdentifier></TargetPlatformIdentifier>
-      <TargetPlatformVersion></TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="Before_ValidateConfiguration" BeforeTargets="_ValidateConfiguration">
-    <PropertyGroup>
-      <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
-      <TargetPlatformVersion>$(TargetPlatformMinVersion)</TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="After_ValidateConfiguration" AfterTargets="_ValidateConfiguration">
-    <PropertyGroup>
-      <TargetPlatformIdentifier></TargetPlatformIdentifier>
-      <TargetPlatformVersion></TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="Before_GetAppxManifestSchemaItemsFromSdk" BeforeTargets="_GetAppxManifestSchemaItemsFromSdk">
-    <PropertyGroup>
-      <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
-      <TargetPlatformVersion>$(TargetPlatformMinVersion)</TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="After_GetAppxManifestSchemaItemsFromSdk" AfterTargets="_GetAppxManifestSchemaItemsFromSdk">
-    <PropertyGroup>
-      <TargetPlatformIdentifier></TargetPlatformIdentifier>
-      <TargetPlatformVersion></TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="Before_GenerateProjectPriConfigurationFiles" BeforeTargets="_GenerateProjectPriConfigurationFiles">
-    <PropertyGroup>
-      <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
-      <TargetPlatformVersion>$(TargetPlatformMinVersion)</TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-  <Target Name="After_GenerateProjectPriConfigurationFiles" AfterTargets="_GenerateProjectPriConfigurationFiles">
-    <PropertyGroup>
-      <TargetPlatformIdentifier></TargetPlatformIdentifier>
-      <TargetPlatformVersion></TargetPlatformVersion>
-    </PropertyGroup>
-  </Target>
-
-</Project>
